Add a non-root user and other refinements
Working as a non-root user makes it harder for a buggy script to
overwrite files that it shouldn't.  N.B. Users will probably need to
wipe out their build directories and start fresh due to this change.

A new "enter" script simplifies logging into the container and
propagates the terminal type and local timezone into it.

When using Terminal on macOS, a frog face appears in the window title.

A couple tips on using macOS as a host have been added to README.md.